pyproject.toml
[project]
name = "shopify-main"
version = "0.1.0"
description = "A FastAPI app for handling Shopify webhooks, customer codes, and REST API with PostgreSQL."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
requires-python = ">=3.9"

[tool.poetry.dependencies]
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
httpx = "^0.28.0"
jinja2 = "^3.1.0"
python-dotenv = "^1.1.0"
pydantic = "^2.11.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.12.0"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"